refactor: implement theme in hook
